feat: Define `set`, `__lshift__` and `__rshift__` methods on `SuperDirtStream`
Tries to make it easier for the user to assign patterns to a stream.

Because `<<` and `>>` are left-associative, they combine the right-side
pattern with the pattern previously set on the stream, using the
corresponding operator defined on `Pattern`, and sets the combination as
the current pattern on the stream.  This allows the user to concatenate
multiple `<<` and `>>` without adding extra parenthesis.

For example:

	p("keys") >> n("0 2 5 6") >> s("superpiano")
	p("drums") >> stack(s("bd*2"), s("hh*8")) >> gain(0.8)

This is the same as:

	p("keys").set(n("0 2 5 6") >> s("superpiano"))
	p("drums").set(stack(s("bd*2"), s("hh*8")) >> gain(0.8))